chore: Get rid of import * utils
