+++ bench/utils/__init__.py
@@ -13,7 +13,7 @@ import click
 
 # imports - module imports
 import bench
-from bench.exceptions import InvalidRemoteException
+from bench.exceptions import InvalidRemoteException, ValidationError
 
 
 logger = logging.getLogger(bench.PROJECT_NAME)
@@ -393,8 +393,22 @@ def is_git_url(url):
 	return bool(re.match(pattern, url))
 
 
-# to avoid circular imports
-from .app import *
-from .bench import *
-from .system import *
-from .translation import *
+def drop_privileges(uid_name='nobody', gid_name='nogroup'):
+	# from http://stackoverflow.com/a/2699996
+	if os.getuid() != 0:
+		# We're not root so, like, whatever dude
+		return
+
+	# Get the uid/gid from the name
+	running_uid = pwd.getpwnam(uid_name).pw_uid
+	running_gid = grp.getgrnam(gid_name).gr_gid
+
+	# Remove group privileges
+	os.setgroups([])
+
+	# Try setting the new uid/gid
+	os.setgid(running_gid)
+	os.setuid(running_uid)
+
+	# Ensure a very conservative umask
+	os.umask(0o22)
