+++ bench/utils/system.py
@@ -135,27 +135,6 @@ def backup_all_sites(bench_path='.'):
 		backup_site(site, bench_path=bench_path)
 
 
-def drop_privileges(uid_name='nobody', gid_name='nogroup'):
-	# from http://stackoverflow.com/a/2699996
-	if os.getuid() != 0:
-		# We're not root so, like, whatever dude
-		return
-
-	# Get the uid/gid from the name
-	running_uid = pwd.getpwnam(uid_name).pw_uid
-	running_gid = grp.getgrnam(gid_name).gr_gid
-
-	# Remove group privileges
-	os.setgroups([])
-
-	# Try setting the new uid/gid
-	os.setgid(running_gid)
-	os.setuid(running_uid)
-
-	# Ensure a very conservative umask
-	os.umask(0o22)
-
-
 def fix_prod_setup_perms(bench_path='.', frappe_user=None):
 	from glob import glob
 	from bench.bench import Bench
